refactor: FeedCard 정리(필요없느 파일 삭제, 코드 나누기)
