--- src/components/FeedCard/FeedTextReadMore/style.ts
@@ -1,10 +1,0 @@
-import styled from '@emotion/styled';
-
-export const ContentsWrapper = styled.div`               
-  overflow: 'hidden',
-  textOverflow: 'ellipsis';
-  overflow: hidden;
-  display: -webkit-box;
-  -webkit-line-clamp: 4;
-  -webkit-box-orient: vertical;
-  `;
